+++ tests/test_helpers/utils.py
@@ -1,5 +1,6 @@
 import importlib.util
 import os
+import subprocess
 
 import pytest
 
@@ -9,6 +10,12 @@ from inspect_ai.solver import TaskState
 
 
 def skip_if_env_var(var: str, exists=True):
+    """
+    Pytest mark to skip the test if the var environment variable is not defined.
+
+    Use in combination with `pytest.mark.api` if the environment variable in
+    question corresponds to a paid API. For example, see `skip_if_no_openai`.
+    """
     condition = (var in os.environ.keys()) if exists else (var not in os.environ.keys())
     return pytest.mark.skipif(
         condition,
@@ -17,7 +24,7 @@ def skip_if_env_var(var: str, exists=True):
 
 
 def skip_if_no_groq(func):
-    return skip_if_env_var("GROQ_API_KEY", exists=False)(func)
+    return pytest.mark.api(skip_if_env_var("GROQ_API_KEY", exists=False)(func))
 
 
 def skip_if_no_package(package):
@@ -35,40 +42,63 @@ def skip_if_no_transformers(func):
     return skip_if_no_package("transformers")(func)
 
 
+def skip_if_no_accelerate(func):
+    return skip_if_no_package("accelerate")(func)
+
+
 def skip_if_no_openai(func):
-    return skip_if_env_var("OPENAI_API_KEY", exists=False)(func)
+    return pytest.mark.api(skip_if_env_var("OPENAI_API_KEY", exists=False)(func))
 
 
 def skip_if_no_anthropic(func):
-    return skip_if_env_var("ANTHROPIC_API_KEY", exists=False)(func)
+    return pytest.mark.api(skip_if_env_var("ANTHROPIC_API_KEY", exists=False)(func))
 
 
 def skip_if_no_google(func):
-    return skip_if_env_var("GOOGLE_API_KEY", exists=False)(func)
+    return pytest.mark.api(skip_if_env_var("GOOGLE_API_KEY", exists=False)(func))
 
 
 def skip_if_no_mistral(func):
-    return skip_if_env_var("MISTRAL_API_KEY", exists=False)(func)
+    return pytest.mark.api(skip_if_env_var("MISTRAL_API_KEY", exists=False)(func))
 
 
 def skip_if_no_cloudflare(func):
-    return skip_if_env_var("CLOUDFLARE_API_TOKEN", exists=False)(func)
+    return pytest.mark.api(skip_if_env_var("CLOUDFLARE_API_TOKEN", exists=False)(func))
 
 
 def skip_if_no_together(func):
-    return skip_if_env_var("TOGETHER_API_KEY", exists=False)(func)
+    return pytest.mark.api(skip_if_env_var("TOGETHER_API_KEY", exists=False)(func))
 
 
 def skip_if_no_azureai(func):
-    return skip_if_env_var("AZURE_API_KEY", exists=False)(func)
+    return pytest.mark.api(skip_if_env_var("AZURE_API_KEY", exists=False)(func))
 
 
 def skip_if_no_vertex(func):
-    return skip_if_env_var("ENABLE_VERTEX_TESTS", exists=False)(func)
+    return pytest.mark.api(skip_if_env_var("ENABLE_VERTEX_TESTS", exists=False)(func))
 
 
 def skip_if_github_action(func):
     return skip_if_env_var("GITHUB_ACTIONS", exists=True)(func)
+
+
+def skip_if_no_docker(func):
+    try:
+        is_docker_installed = (
+            subprocess.run(
+                ["docker", "--version"],
+                check=False,
+                stdout=subprocess.PIPE,
+                stderr=subprocess.PIPE,
+            ).returncode
+            == 0
+        )
+    except FileNotFoundError:
+        is_docker_installed = False
+
+    return pytest.mark.skipif(
+        not is_docker_installed, reason="Test doesn't work without Docker installed."
+    )(func)
 
 
 def run_example(example: str, model: str):